requirements.txt
Flask==2.3.3
requests==2.26.0
openai==0.27.0
python-dotenv==1.0.0
marshmallow==3.20.1